requirements.txt
datasets
nltk
sentence-transformers
faiss-cpu
evaluate
requests
numpy